+++ webRTCAPP/Classes/webrtc/modules/audio_coding/neteq/tools/neteq_quality_test.h
@@ -38,9 +38,10 @@ class NoLoss : public LossModel {
 
 class UniformLoss : public LossModel {
  public:
-  UniformLoss(int loss_rate);
+  UniformLoss(double loss_rate);
   virtual bool Lost() OVERRIDE;
   void set_loss_rate(double loss_rate) { loss_rate_ = loss_rate; }
+
  private:
   double loss_rate_;
 };
@@ -49,6 +50,7 @@ class GilbertElliotLoss : public LossModel {
  public:
   GilbertElliotLoss(double prob_trans_11, double prob_trans_01);
   virtual bool Lost() OVERRIDE;
+
  private:
   // Prob. of losing current packet, when previous packet is lost.
   double prob_trans_11_;
@@ -116,7 +118,7 @@ class NetEqQualityTest : public ::testing::Test {
   // Expected output number of samples per channel in a frame.
   const int out_size_samples_;
 
-  int payload_size_bytes_;
+  size_t payload_size_bytes_;
   int max_payload_bytes_;
 
   scoped_ptr<InputAudioFile> in_file_;
@@ -132,7 +134,7 @@ class NetEqQualityTest : public ::testing::Test {
   scoped_ptr<int16_t[]> out_data_;
   WebRtcRTPHeader rtp_header_;
 
-  long total_payload_size_bytes_;
+  size_t total_payload_size_bytes_;
 };
 
 }  // namespace test
